+++ .storybook/webpack.config.js
@@ -27,6 +27,12 @@ module.exports = {
                 test: /\.svg$/,
                 loader: 'svg-url-loader',
             },
+            {
+                test: /\.(jsx?|tsx?)$/,
+                // Exclude all node_modules from transpilation, except for Ariakit dependencies
+                exclude: /node_modules\/(?!(ariakit.*)\/).*/,
+                use: [{ loader: 'babel-loader' }],
+            },
         ],
     },
     resolve: {